Fix poll count in ibv_transmit::wait_for_wc
It was waiting for up to batch_size completions, but each completion is
for a whole batch, so in almost all cases a single completion is
sufficient. The code wasn't wrong, but might have led to some unevenness
in packet pacing since after a delay there would be more completions to
reap and hence more delay due to reaping them.